+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -114,6 +114,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added a Lighting and BackLighting slots in Lit, StackLit, Fabric and Hair master nodes
 - Added xrInstancing flag support to RTHandle
 - Added support for cullmask for decal projectors
+- Added emissive property for shader graph decals
 
 ### Fixed
 - Fixed logic to disable FPTL with stereo rendering